requirements.txt
asgiref==3.3.1
Django==3.1.4
django-cors-headers==3.6.0
djangorestframework==3.12.2
djangorestframework-simplejwt==4.6.0
Pillow==8.0.1
psycopg2-binary==2.8.6
PyJWT==1.7.1
pytz==2020.4
sqlparse==0.4.1
asgiref==3.3.1
certifi==2020.12.5
chardet==4.0.0
Django==3.1.4
django-cors-headers==3.6.0
djangorestframework==3.12.2
djangorestframework-simplejwt==4.6.0
idna==2.10
Pillow==8.0.1
psycopg2-binary==2.8.6
PyJWT==1.7.1
pytz==2020.4
requests==2.25.1
sqlparse==0.4.1
urllib3==1.26.2
amqp==5.0.2
asgiref==3.3.1
billiard==3.6.3.0
celery==4.4.7
certifi==2020.12.5
chardet==4.0.0
click==7.1.2
click-didyoumean==0.0.3
click-plugins==1.1.1
click-repl==0.1.6
Django==3.1.4
django-cors-headers==3.6.0
djangorestframework==3.12.2
djangorestframework-simplejwt==4.6.0
idna==2.10
kombu==5.0.2
Pillow==8.0.1
prompt-toolkit==3.0.8
psycopg2-binary==2.8.6
PyJWT==1.7.1
pytz==2020.4
requests==2.25.1
six==1.15.0
sqlparse==0.4.1
urllib3==1.26.2
vine==5.0.0
wcwidth==0.2.5
amqp==5.0.2
asgiref==3.3.1
billiard==3.6.3.0
celery==5.0.5
certifi==2020.12.5
chardet==4.0.0
click==7.1.2
click-didyoumean==0.0.3
click-plugins==1.1.1
click-repl==0.1.6
Django==3.1.4
django-cors-headers==3.6.0
djangorestframework==3.12.2
djangorestframework-simplejwt==4.6.0
idna==2.10
kombu==5.0.2
Pillow==8.0.1
prompt-toolkit==3.0.8
psycopg2-binary==2.8.6
PyJWT==1.7.1
pytz==2020.4
redis==3.5.3
requests==2.25.1
six==1.15.0
sqlparse==0.4.1
urllib3==1.26.2
vine==5.0.0
wcwidth==0.2.5
beautifulsoup4==4.9.3
termcolor==1.1.0
channels==3.0.3
coloredlogs==15.0